What the Fuck? Fixing Issues Causing Another Issue
I backtracked the code manually, removing lines I believed to be causing
the issue. I pretty much reversed most of the integration of the
Variables.vb file into the movement test directories, since after that
commit is when I experienced the issues. The movement issue was not
solved, so I moved on to the form transition issue. Simply creating a
new main form solved the transitioning.  I played around with the
designer for the MovementFormNew and disabling the button on it allowed
me to make movement again using the arrow keys. The issue is temporarily
fixed and testing can continue, but I can't put it off for long.