New Saturday commits
Added on to the Choice form when you complete the rebuilding of the
city.

Made the textbox in the dev console to a combo box to eliminate the
possibility of most errors.

Added labels for QA testers on the debugging value labels in the idle
clicker game

Made the "Do not ask again" option function using a variable boolean on
the recruit civilians button

Other little stuff.