3-22-16 Gen Progress
BrillStress > The generators are done in full. It will calculate the
Class, Armor, Weapon, and Number of said enemy (For town
invasions/Raids) while generating a possible 5v1 fight versus the
player.

WepArmTest > Didn't do much with this today, but I will be adding a
healing rng for the enemy as well as the player tomorrow. The progress
bar value fix (listed in the Issues) comes from this program.